Write AI controller

+++ backend/src/main/java/com/dp/backend/controller/ReviewController.java
@@ -1,6 +1,7 @@
 package com.dp.backend.controller;
 
 import com.dp.backend.model.Review;
+import com.dp.backend.repository.ReviewRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -15,7 +16,7 @@ public class ReviewController {
     ReviewRepository reviewRepository;
 
     @GetMapping("/{id}")
-    public List<Review> getReviewsByProduct(@PathVariable Long id) {
+    public List<Review> getReviewsByProduct(@PathVariable int id) {
         return reviewRepository.findByProductId(id);
     }
 }
